refactor(refs DPLAN-4633): extract duplicated methods into AssessmentTableFormattingTrait (5728)
* feat(refs DPLAN-16641): add votePla support to PDF exports

 - Add votePla field to formatStatementArray in AssessmentTableServiceOutput
 - Inject GlobalConfigInterface into AssessmentTablePdfExporter to access form_options
 - Translate votePla values to readable text before rendering PDF templates
 - Update all 5 PDF LaTeX templates to display votePla text before recommendation:
 - export_assessment_table_entry.tex.twig
 - export_assessment_table_entry_condensed.tex.twig (both branches)
 - export_assessment_table_entry_anonymous.tex.twig
 - export_assessment_table_entry_anonymous_condensed.tex.twig
 - export_fragments_anonymous.tex.twig
 - votePla text appears with paragraph break (\n\n) before recommendation text The Abwägungsvorschlag dropdown value is now included in all PDF export formats (regular, condensed, anonymous).

* feat(refs DPLAN-16641): add votePla support to DOCX compact export

  Add Abwägungsvorschlag dropdown value display in DOCX "Querformat (kompakt)" export format.

  Changes:
  - Include votePla field in statement data array (formatStatementArray)
  - Display votePla text before recommendation in compact export (renderTableItem)
  - Apply permission check (field_vote_advice_docx) for visibility
  - Translate votePla keys using statement_fragment_advice_values configuration

* refactor(refs DPLAN-16641): reduce cognitive complexity in DocxExporter

  Extract votePla display logic into separate method to reduce cognitive complexity of renderTableItem from 22 to under the allowed limit of 15.

  Changes:
  - Extract addVotePlaText() private method from renderTableItem()
  - Use early return pattern to reduce nesting
  - Maintain same functionality with improved code organization

* refactor(refs DPLAN-4633): extract duplicated methods into AssessmentTableFormattingTrait

  Extract common formatting methods from AssessmentTableServiceOutput and DocxExporter into a shared trait to eliminate code duplication reported by SonarCloud.

  Changes:
  - Create AssessmentTableFormattingTrait with shared methods
  - Add trait to both AssessmentTableServiceOutput and DocxExporter
  - Remove duplicated methods from both classes:
    - getDefaultDocxPageStyles()
    - getDefaultDocxTableStyle()
    - getDocxImageTag()
    - formatStatementArray()
    - formatStatementObject()
    - formatFragmentArray()
    - renderGroup()
    - createExternIdString()
    - createExternIdStringFromObject()
  - Add getLogger() method to AssessmentTableServiceOutput
  - Define abstract getLogger() method in trait for logger access

* refactor(refs DPLAN-4633): move createFormerProcedureSuffix into AssessmentTableFormattingTrait

The identical private helper was duplicated in both consumers while the
trait already depended on it. Centralising it removes the duplication and
the trait's hidden dependency on each consumer providing the method.

* feat(refs DPLAN-4633): Remove unused injection.

* style: Apply php-cs-fixer

---------
