+++ demosplan/DemosPlanCoreBundle/Logic/AssessmentTable/AssessmentTableServiceOutput.php
@@ -20,11 +20,10 @@ use demosplan\DemosPlanCoreBundle\Entity\StatementAttachment;
 use demosplan\DemosPlanCoreBundle\Exception\AccessDeniedGuestException;
 use demosplan\DemosPlanCoreBundle\Exception\MessageBagException;
 use demosplan\DemosPlanCoreBundle\Exception\StatementElementNotFoundException;
-use demosplan\DemosPlanCoreBundle\Logic\Export\DocumentWriterSelector;
+use demosplan\DemosPlanCoreBundle\Logic\Export\AssessmentTableFormattingTrait;
 use demosplan\DemosPlanCoreBundle\Logic\Export\DocxExporter;
 use demosplan\DemosPlanCoreBundle\Logic\Export\PhpWordConfigurator;
 use demosplan\DemosPlanCoreBundle\Logic\FileService;
-use demosplan\DemosPlanCoreBundle\Logic\Grouping\StatementEntityGroup;
 use demosplan\DemosPlanCoreBundle\Logic\Map\MapService;
 use demosplan\DemosPlanCoreBundle\Logic\MessageBag;
 use demosplan\DemosPlanCoreBundle\Logic\Procedure\ProcedureHandler;
@@ -42,14 +41,12 @@ use Exception;
 use Illuminate\Support\Collection;
 use League\Flysystem\FilesystemOperator;
 use PhpOffice\PhpWord\Element\AbstractContainer;
-use PhpOffice\PhpWord\Element\Section;
 use PhpOffice\PhpWord\Element\Table;
 use PhpOffice\PhpWord\PhpWord;
 use PhpOffice\PhpWord\Shared\Html;
 use PhpOffice\PhpWord\SimpleType\Jc;
 use PhpOffice\PhpWord\Writer\WriterInterface;
 use Psr\Log\LoggerInterface;
-use ReflectionException;
 use Symfony\Component\Filesystem\Filesystem;
 use Symfony\Component\Form\FormFactory;
 use Symfony\Component\Form\FormFactoryInterface;
@@ -69,6 +66,8 @@ use function substr;
 
 class AssessmentTableServiceOutput
 {
+    use AssessmentTableFormattingTrait;
+
     final public const EXPORT_SORT_BY_PARAGRAPH_FRAGMENTS_ONLY = 'byParagraphFragmentsOnly';
     final public const EXPORT_SORT_BY_PARAGRAPH = 'byParagraph';
     final public const EXPORT_SORT_DEFAULT = 'default';
@@ -148,7 +147,6 @@ class AssessmentTableServiceOutput
         StatementService $statementService,
         private readonly TranslatorInterface $translator,
         ValidatorInterface $validator,
-        private readonly DocumentWriterSelector $writerSelector,
     ) {
         $this->assessmentTableServiceStorage = $assessmentTableServiceStorage;
         $this->config = $config;
@@ -607,51 +605,9 @@ class AssessmentTableServiceOutput
         );
     }
 
-    /**
-     * Get default Docx Page Styles.
-     */
-    protected function getDefaultDocxPageStyles(ViewOrientation $orientation): array
+    protected function getLogger(): LoggerInterface
     {
-        $styles = [];
-        // Benutze das ausgewählte Format
-        $styles['orientation'] = [];
-        // im Hochformat werden für LibreOffice anderen Breiten benötigt
-        $styles['cellWidthTotal'] = 10000;
-        $styles['firstCellWidth'] = 1500;
-        $styles['cellWidth'] = 3850;
-        $styles['cellWidthSecondThird'] = 7500;
-
-        $tableStyle = $this->getDefaultDocxTableStyle();
-        $styles['tableStyle'] = $tableStyle;
-        $styles['cellStyleStatementDetails'] = ['gridSpan' => 2, 'bgColor' => 'f0f0f5', 'valign' => 'top'];
-        $styles['textStyleStatementDetails'] = ['bold' => true];
-        $styles['textStyleStatementDetailsParagraphStyles'] = ['spaceAfter' => 0];
-        $styles['cellHeading'] = ['align' => 'center', 'valign' => 'center'];
-        $styles['cellHeadingText'] = ['bold' => true, 'valign' => 'center', 'align' => 'center', 'name' => 'Arial', 'size' => 9];
-        $styles['cellTop'] = ['valign' => 'top'];
-
-        if ($orientation->isLandscape()) {
-            $styles['cellWidthTotal'] = 14000;
-            $styles['orientation'] = ['orientation' => 'landscape'];
-            $styles['firstCellWidth'] = 2000;
-            $styles['cellWidth'] = 6000;
-            $styles['cellWidthSecondThird'] = 12000;
-        }
-
-        return $styles;
-    }
-
-    protected function getDefaultDocxTableStyle(): \PhpOffice\PhpWord\Style\Table
-    {
-        $tableStyle = new \PhpOffice\PhpWord\Style\Table();
-        $formatSpecificTableStyle = $this->writerSelector->getTableStyleForFormat($this->tableStyle);
-
-        $tableStyle->setLayout(\PhpOffice\PhpWord\Style\Table::LAYOUT_FIXED)
-            ->setBorderColor($formatSpecificTableStyle['borderColor'])
-            ->setBorderSize($formatSpecificTableStyle['borderSize'])
-            ->setCellMargin($formatSpecificTableStyle['cellMargin']);
-
-        return $tableStyle;
+        return $this->logger;
     }
 
     /**
@@ -726,80 +682,6 @@ class AssessmentTableServiceOutput
     }
 
     /**
-     * Generate Html imagetag to be used in PhphWord Html::addHtml().
-     *
-     * @param string $imageFile
-     * @param int    $maxWidth  maximum image width in pixel
-     *
-     * @return string
-     */
-    protected function getDocxImageTag($imageFile, $maxWidth = 500)
-    {
-        $imgTag = '';
-        $width = 300;
-        $height = 300;
-        $margin = 10;
-        // phpword needs a local file, no need for flysystem
-        if (!file_exists($imageFile)) {
-            return $imgTag;
-        }
-
-        // get Image size
-        $imageInfo = getimagesize($imageFile);
-        if (2 < (is_countable($imageInfo) ? count($imageInfo) : 0)) {
-            $width = $imageInfo[0] - $margin;
-            $height = $imageInfo[1] - $margin;
-        }
-
-        // check that picture is not wider than allowed
-        if ($width > $maxWidth) {
-            $factor = $width / $maxWidth;
-
-            // resize Image
-            if (0 != $factor) {
-                $width /= $factor;
-                $height /= $factor;
-            }
-            $this->logger->info('Docx Image resize to width: '.$width.' and height: '.$height);
-        }
-
-        return '<img height="'.$height.'" width="'.$width.'" src="'.$imageFile.'"/>';
-    }
-
-    /**
-     * Statement in unified data format.
-     *
-     * @return array - formatted statement
-     *
-     * @throws ReflectionException
-     *
-     * @deprecated Use {@link formatStatementObject} instead
-     */
-    public function formatStatementArray(array $statement): array
-    {
-        $formatted = $this->docxExporter->formatStatementArray($statement);
-        $formatted['isSubmittedByCitizen'] = $statement['isSubmittedByCitizen'] ?? null;
-
-        return $formatted;
-    }
-
-    /**
-     * Statement in unified data format.
-     *
-     * @return array formatted statement
-     *
-     * @throws ReflectionException
-     */
-    public function formatStatementObject(Statement $statement): array
-    {
-        $item = $this->statementService->convertToLegacy($statement);
-        $item['parent'] = $this->statementService->convertToLegacy($statement->getParent());
-        $item['original'] = $this->statementService->convertToLegacy($statement->getOriginal());
-
-        return $this->formatStatementArray($item);
-    }
-
-    /**
      * Fragment in unified data format.
      *
      * @return array - formatted fragment
@@ -817,45 +699,6 @@ class AssessmentTableServiceOutput
         $item['text'] = $fragment->getText();
         $item['elementId'] = $fragment->getElementId();
         $item['elementTitle'] = $fragment->getElementTitle();
-
-        return $item;
-    }
-
-    /**
-     * Fragment in unified data format.
-     *
-     * @return array - formatted fragment
-     *
-     * @throws Exception
-     *
-     * @deprecated Use {@link formatFragmentObject} instead
-     */
-    public function formatFragmentArray(array $statement, array $fragment): array
-    {
-        $tmpElementId = $fragment['elementId'];
-        $tmpElementTitle = $fragment['elementTitle'];
-
-        $item = $this->formatStatementArray($statement);
-        $item['sortIndex'] = $fragment['sortIndex'];
-
-        // override selected item fields with fragment content:
-        $item['type'] = 'fragment';
-        $item['created'] = $fragment['created'] ?? null;
-
-        $item['text'] = '';
-        $item['recommendation'] = '';
-        // we need to fetch Fragment, as text fields are not mapped in statement
-        // index for performance reasons
-        $statementFragment = $this->statementHandler->getStatementFragment($fragment['id']);
-        if ($statementFragment instanceof StatementFragment) {
-            // pretend as if consideration would be an recommendation
-            // as it has the same behaviour
-            $item['recommendation'] = $statementFragment->getConsideration();
-            $item['text'] = $statementFragment->getText();
-        }
-
-        $item['elementId'] = $tmpElementId;
-        $item['elementTitle'] = $tmpElementTitle;
 
         return $item;
     }
@@ -1012,114 +855,6 @@ class AssessmentTableServiceOutput
             $table->getPager(),
             $table->getNavigation()
         );
-    }
-
-    /**
-     * @throws Exception
-     */
-    protected function renderGroup(
-        StatementEntityGroup $group,
-        callable $entriesRenderFunction,
-        Section $section,
-        int $depth = 0,
-    ): void {
-        $section->addTitle($group->getTitle(), $depth + 2);
-
-        foreach ($group->getSubgroups() as $subgroup) {
-            $this->renderGroup(
-                $subgroup,
-                $entriesRenderFunction,
-                $section,
-                $depth + 1
-            );
-        }
-
-        if (0 !== (is_countable($group->getEntries()) ? count($group->getEntries()) : 0)) {
-            $entriesRenderFunction($section, $group->getEntries());
-        }
-    }
-
-    /**
-     * T10049
-     * Centralisation of logic to generate string of externId.
-     *
-     * Includes "Kopie von" in case of current statement is a copy of a statement and placeholder statement information.
-     *
-     * @param array $statementArray
-     *
-     * @deprecated use {@link AssessmentTableServiceOutput::createExternIdStringFromObject} instead
-     */
-    public function createExternIdString($statementArray): string
-    {
-        $externIdString = '';
-
-        // add "copyof"
-        if (isset($statementArray['originalId']) && isset($statementArray['parentId'])
-            && $statementArray['originalId'] != $statementArray['parentId']
-            && false === is_null($statementArray['parentId'])) {
-            $externIdString .= $this->translator->trans('copyof').' ';
-        }
-
-        $externIdString .= $statementArray['externId'];
-
-        // add former externID in case of statement was moved from another procedure
-
-        // was moved?
-        if (array_key_exists('formerExternId', $statementArray) && false === is_null($statementArray['formerExternId'])) {
-            $externIdString .= ' ('.$this->translator->trans('formerExternId').': '.$statementArray['formerExternId'].' '.$this->translator->trans('from').' '.$statementArray['movedFromProcedureName'].')';
-        } elseif (array_key_exists('placeholderStatement', $statementArray)
-            && false === is_null($statementArray['placeholderStatement'])) {
-            // dont know, if $statementArray['placeholderStatement'] is an object or array. -> handle both cases:
-            if ($statementArray['placeholderStatement'] instanceof Statement) {
-                $formerExternId = $statementArray['placeholderStatement']->getExternId();
-                $nameOfFormerProcedure = $statementArray['placeholderStatement']->getProcedure()->getName();
-            } else {
-                $formerExternId = $statementArray['placeholderStatement']['externId'];
-                $nameOfFormerProcedure = $statementArray['placeholderStatement']['procedure']['name'];
-            }
-            $externIdString .= ' ('.$this->translator->trans('formerExternId').': '.$formerExternId.' '.$this->translator->trans('from').' '.$nameOfFormerProcedure.')';
-        }
-
-        // if statement was moved into another procedure, this will usually be displayed in the textfield of the statement
-        return $externIdString;
-    }
-
-    /**
-     * T10049
-     * Centralisation of logic to generate string of externId.
-     *
-     * Includes "Kopie von" in case of current statement is a copy of a statement and placeholder statement information.
-     */
-    public function createExternIdStringFromObject(Statement $statement): string
-    {
-        $externIdString = $statement->getExternId();
-
-        // add "copyof"
-        if (null !== $statement->getParentId()
-            && $statement->getOriginalId() != $statement->getParentId()) {
-            $externIdString = $this->translator->trans('copyof').' '.$externIdString;
-        }
-
-        // add former externID in case of statement was moved from another procedure
-        // was moved?
-        $placeholderStatement = $statement->getPlaceholderStatement();
-        if (null !== $statement->getFormerExternId()) {
-            $formerExternId = $statement->getFormerExternId();
-            $nameOfFormerProcedure = $statement->getMovedFromProcedureName();
-            $externIdString .= $this->createFormerProcedureSuffix($formerExternId, $nameOfFormerProcedure);
-        } elseif (null !== $placeholderStatement) {
-            $formerExternId = $placeholderStatement->getExternId();
-            $nameOfFormerProcedure = $placeholderStatement->getProcedure()->getName();
-            $externIdString .= $this->createFormerProcedureSuffix($formerExternId, $nameOfFormerProcedure);
-        }
-
-        // if statement was moved into another procedure, this will usually be displayed in the textfield of the statement
-        return $externIdString;
-    }
-
-    private function createFormerProcedureSuffix($formerExternId, $nameOfFormerProcedure): string
-    {
-        return ' ('.$this->translator->trans('formerExternId').': '.$formerExternId.' '.$this->translator->trans('from').' '.$nameOfFormerProcedure.')';
     }
 
     /**
