+++ demosplan/DemosPlanCoreBundle/Logic/Export/DocxExporter.php
@@ -21,7 +21,6 @@ use demosplan\DemosPlanCoreBundle\Entity\Statement\Statement;
 use demosplan\DemosPlanCoreBundle\Entity\Statement\StatementFragment;
 use demosplan\DemosPlanCoreBundle\Entity\StatementAttachment;
 use demosplan\DemosPlanCoreBundle\Exception\InvalidArgumentException;
-use demosplan\DemosPlanCoreBundle\Logic\AssessmentTable\AssessmentTableServiceOutput;
 use demosplan\DemosPlanCoreBundle\Logic\AssessmentTable\AssessmentTableViewMode;
 use demosplan\DemosPlanCoreBundle\Logic\AssessmentTable\ViewOrientation;
 use demosplan\DemosPlanCoreBundle\Logic\EditorService;
@@ -40,7 +39,6 @@ use demosplan\DemosPlanCoreBundle\ValueObject\AssessmentTable\StatementHandlingR
 use Exception;
 use Illuminate\Support\Collection;
 use League\Flysystem\FilesystemOperator;
-use Monolog\Logger;
 use PhpOffice\PhpWord\Element\AbstractContainer;
 use PhpOffice\PhpWord\Element\Cell;
 use PhpOffice\PhpWord\Element\Footer;
@@ -50,7 +48,6 @@ use PhpOffice\PhpWord\PhpWord;
 use PhpOffice\PhpWord\Shared\Html;
 use PhpOffice\PhpWord\Writer\WriterInterface;
 use Psr\Log\LoggerInterface;
-use ReflectionException;
 use Symfony\Component\Filesystem\Filesystem;
 use Symfony\Component\Form\FormFactory;
 use Symfony\Component\Validator\Validator\ValidatorInterface;
@@ -60,6 +57,7 @@ use Twig\Environment;
 class DocxExporter
 {
     use RequiresTranslatorTrait;
+    use AssessmentTableFormattingTrait;
 
     final public const EXPORT_SORT_BY_PARAGRAPH_FRAGMENTS_ONLY = 'byParagraphFragmentsOnly';
     final public const EXPORT_SORT_BY_PARAGRAPH = 'byParagraph';
@@ -488,52 +486,6 @@ class DocxExporter
             ->all();
     }
 
-    /**
-     * Get default Docx Page Styles.
-     */
-    protected function getDefaultDocxPageStyles(ViewOrientation $orientation): array
-    {
-        $styles = [];
-        // Benutze das ausgewählte Format
-        $styles['orientation'] = [];
-        // im Hochformat werden für LibreOffice anderen Breiten benötigt
-        $styles['cellWidthTotal'] = 10000;
-        $styles['firstCellWidth'] = 1500;
-        $styles['cellWidth'] = 3850;
-        $styles['cellWidthSecondThird'] = 7500;
-
-        $tableStyle = $this->getDefaultDocxTableStyle();
-        $styles['tableStyle'] = $tableStyle;
-        $styles['cellStyleStatementDetails'] = ['gridSpan' => 2, 'bgColor' => 'f0f0f5', 'valign' => 'top'];
-        $styles['textStyleStatementDetails'] = ['bold' => true];
-        $styles['textStyleStatementDetailsParagraphStyles'] = ['spaceAfter' => 0];
-        $styles['cellHeading'] = ['align' => 'center', 'valign' => 'center'];
-        $styles['cellHeadingText'] = ['bold' => true, 'valign' => 'center', 'align' => 'center', 'name' => 'Arial', 'size' => 9];
-        $styles['cellTop'] = ['valign' => 'top'];
-
-        if ($orientation->isLandscape()) {
-            $styles['cellWidthTotal'] = 14000;
-            $styles['orientation'] = ['orientation' => 'landscape'];
-            $styles['firstCellWidth'] = 2000;
-            $styles['cellWidth'] = 6000;
-            $styles['cellWidthSecondThird'] = 12000;
-        }
-
-        return $styles;
-    }
-
-    protected function getDefaultDocxTableStyle(): \PhpOffice\PhpWord\Style\Table
-    {
-        $tableStyle = new \PhpOffice\PhpWord\Style\Table();
-
-        $tableStyle->setLayout(\PhpOffice\PhpWord\Style\Table::LAYOUT_FIXED);
-        $tableStyle->setBorderColor($this->tableStyle['borderColor']);
-        $tableStyle->setBorderSize($this->tableStyle['borderSize']);
-        $tableStyle->setCellMargin($this->tableStyle['cellMargin']);
-
-        return $tableStyle;
-    }
-
     public function addCondensedTableHeaders(array $styles, Table $table, string $typeHeader): void
     {
         $table->addRow(null, ['tblHeader' => true]);
@@ -566,75 +518,6 @@ class DocxExporter
         $item['elementTitle'] = $fragment->getElementTitle();
 
         return $item;
-    }
-
-    /**
-     * Statement in unified data format.
-     *
-     * @return array - formatted statement
-     *
-     * @throws ReflectionException
-     *
-     * @deprecated Use {@link formatStatementObject} instead
-     */
-    public function formatStatementArray(array $statement): array
-    {
-        return [
-            'type'                      => 'statement',
-            'attachments'               => $statement['attachments'] ?? null,
-            'authoredDate'              => $statement['meta']['authoredDate'] ?? null,
-            'cluster'                   => $statement['cluster'] ?? null,
-            'documentTitle'             => $statement['document']['title'] ?? null,
-            'externId'                  => $statement['externId'] ?? null,
-            'formerExternId'            => $statement['formerExternId'] ?? null,
-            'elementTitle'              => $statement['element']['title'] ?? null,
-            'files'                     => $statement['files'] ?? null,
-            'orgaName'                  => $statement['meta']['orgaName'] ?? null,
-            'orgaDepartmentName'        => $statement['meta']['orgaDepartmentName'] ?? null,
-            'originalId'                => $statement['original']['ident'] ?? null,
-            'paragraphTitle'            => $statement['paragraph']['title'] ?? null,
-            'parentId'                  => $statement['parent']['ident'] ?? null,
-            'polygon'                   => $statement['polygon'] ?? null,
-            'publicAllowed'             => $statement['publicAllowed'] ?? null,
-            'publicCheck'               => $statement['publicCheck'] ?? null,
-            'publicStatement'           => $statement['publicStatement'] ?? null,
-            'publicVerified'            => $statement['publicVerified'] ?? null,
-            'publicVerifiedTranslation' => $statement['publicVerifiedTranslation'] ?? null,
-            'recommendation'            => $statement['recommendation'] ?? null,
-            'votePla'                   => $statement['votePla'] ?? null,
-            'submit'                    => $statement['submit'] ?? null,
-            'submitName'                => $statement['meta']['submitName'] ?? null,
-            'authorName'                => $statement['meta']['authorName'] ?? null,
-            'text'                      => $statement['text'] ?? null,
-            'votes'                     => $statement['votes'] ?? null,
-            'votesNum'                  => $statement['votesNum'] ?? null,
-            'likesNum'                  => $statement['likesNum'] ?? null,
-            'fragments'                 => [],
-            'userState'                 => $statement['meta']['userState'] ?? null,
-            'userGroup'                 => $statement['meta']['userGroup'] ?? null,
-            'userOrganisation'          => $statement['meta']['userOrganisation'] ?? null,
-            'movedToProcedureName'      => $statement['movedToProcedureName'] ?? null,
-            'movedFromProcedureName'    => $statement['movedFromProcedureName'] ?? null,
-            'userPosition'              => $statement['meta']['userPosition'] ?? null,
-            'isClusterStatement'        => $statement['isClusterStatement'] ?? null,
-            'name'                      => $statement['name'] ?? null,
-        ];
-    }
-
-    /**
-     * Statement in unified data format.
-     *
-     * @return array formatted statement
-     *
-     * @throws ReflectionException
-     */
-    public function formatStatementObject(Statement $statement): array
-    {
-        $item = $this->statementService->convertToLegacy($statement);
-        $item['parent'] = $this->statementService->convertToLegacy($statement->getParent());
-        $item['original'] = $this->statementService->convertToLegacy($statement->getOriginal());
-
-        return $this->formatStatementArray($item);
     }
 
     /**
@@ -851,89 +734,6 @@ class DocxExporter
     }
 
     /**
-     * T10049
-     * Centralisation of logic to generate string of externId.
-     *
-     * Includes "Kopie von" in case of current statement is a copy of a statement and placeholder statement information.
-     */
-    public function createExternIdStringFromObject(Statement $statement): string
-    {
-        $externIdString = $statement->getExternId();
-
-        // add "copyof"
-        if (null !== $statement->getParentId()
-            && $statement->getOriginalId() != $statement->getParentId()) {
-            $externIdString = $this->translator->trans('copyof').' '.$externIdString;
-        }
-
-        // add former externID in case of statement was moved from another procedure
-        // was moved?
-        $placeholderStatement = $statement->getPlaceholderStatement();
-        if (null !== $statement->getFormerExternId()) {
-            $formerExternId = $statement->getFormerExternId();
-            $nameOfFormerProcedure = $statement->getMovedFromProcedureName();
-            $externIdString .= $this->createFormerProcedureSuffix($formerExternId, $nameOfFormerProcedure);
-        } elseif (null !== $placeholderStatement) {
-            $formerExternId = $placeholderStatement->getExternId();
-            $nameOfFormerProcedure = $placeholderStatement->getProcedure()->getName();
-            $externIdString .= $this->createFormerProcedureSuffix($formerExternId, $nameOfFormerProcedure);
-        }
-
-        // if statement was moved into another procedure, this will usually be displayed in the textfield of the statement
-        return $externIdString;
-    }
-
-    /**
-     * T10049
-     * Centralisation of logic to generate string of externId.
-     *
-     * Includes "Kopie von" in case of current statement is a copy of a statement and placeholder statement information.
-     *
-     * @param array $statementArray
-     *
-     * @deprecated use {@link AssessmentTableServiceOutput::createExternIdStringFromObject} instead
-     */
-    public function createExternIdString($statementArray): string
-    {
-        $externIdString = '';
-
-        // add "copyof"
-        if (isset($statementArray['originalId']) && isset($statementArray['parentId'])
-            && $statementArray['originalId'] != $statementArray['parentId']
-            && false === is_null($statementArray['parentId'])) {
-            $externIdString .= $this->translator->trans('copyof').' ';
-        }
-
-        $externIdString .= $statementArray['externId'];
-
-        // add former externID in case of statement was moved from another procedure
-
-        // was moved?
-        if (array_key_exists('formerExternId', $statementArray) && false === is_null($statementArray['formerExternId'])) {
-            $externIdString .= ' ('.$this->translator->trans('formerExternId').': '.$statementArray['formerExternId'].' '.$this->translator->trans('from').' '.$statementArray['movedFromProcedureName'].')';
-        } elseif (array_key_exists('placeholderStatement', $statementArray)
-            && false === is_null($statementArray['placeholderStatement'])) {
-            // dont know, if $statementArray['placeholderStatement'] is an object or array. -> handle both cases:
-            if ($statementArray['placeholderStatement'] instanceof Statement) {
-                $formerExternId = $statementArray['placeholderStatement']->getExternId();
-                $nameOfFormerProcedure = $statementArray['placeholderStatement']->getProcedure()->getName();
-            } else {
-                $formerExternId = $statementArray['placeholderStatement']['externId'];
-                $nameOfFormerProcedure = $statementArray['placeholderStatement']['procedure']['name'];
-            }
-            $externIdString .= ' ('.$this->translator->trans('formerExternId').': '.$formerExternId.' '.$this->translator->trans('from').' '.$nameOfFormerProcedure.')';
-        }
-
-        // if statement was moved into another procedure, this will usually be displayed in the textfield of the statement
-        return $externIdString;
-    }
-
-    private function createFormerProcedureSuffix($formerExternId, $nameOfFormerProcedure): string
-    {
-        return ' ('.$this->translator->trans('formerExternId').': '.$formerExternId.' '.$this->translator->trans('from').' '.$nameOfFormerProcedure.')';
-    }
-
-    /**
      * Add each fragments in $item['fragments'] as a row with the fragment text on the left side and the fragment recommendation
      * on the right side. $item is needed for the submitterData of the first fragment.
      *
@@ -1036,10 +836,7 @@ class DocxExporter
         return preg_replace(array_keys($replacements), array_values($replacements), $text);
     }
 
-    /**
-     * @return Logger
-     */
-    protected function getLogger()
+    protected function getLogger(): LoggerInterface
     {
         return $this->logger;
     }
@@ -1072,31 +869,6 @@ class DocxExporter
         $frontPageSection->addText($translator->trans('procedure.agency').': '.$procedure->getOrgaName(), $coverStyle, $coverParagraphStyle);
 
         return $frontPageSection;
-    }
-
-    /**
-     * @throws Exception
-     */
-    protected function renderGroup(
-        StatementEntityGroup $group,
-        callable $entriesRenderFunction,
-        Section $section,
-        int $depth = 0,
-    ): void {
-        $section->addTitle($group->getTitle(), $depth + 2);
-
-        foreach ($group->getSubgroups() as $subgroup) {
-            $this->renderGroup(
-                $subgroup,
-                $entriesRenderFunction,
-                $section,
-                $depth + 1
-            );
-        }
-
-        if (0 !== (is_countable($group->getEntries()) ? count($group->getEntries()) : 0)) {
-            $entriesRenderFunction($section, $group->getEntries());
-        }
     }
 
     protected function getExportPageHeader(Procedure $procedure): string
@@ -1144,45 +916,6 @@ class DocxExporter
 
                 return $item;
             });
-    }
-
-    /**
-     * Fragment in unified data format.
-     *
-     * @return array - formatted fragment
-     *
-     * @throws Exception
-     *
-     * @deprecated Use {@link formatFragmentObject} instead
-     */
-    public function formatFragmentArray(array $statement, array $fragment): array
-    {
-        $tmpElementId = $fragment['elementId'];
-        $tmpElementTitle = $fragment['elementTitle'];
-
-        $item = $this->formatStatementArray($statement);
-        $item['sortIndex'] = $fragment['sortIndex'];
-
-        // override selected item fields with fragment content:
-        $item['type'] = 'fragment';
-        $item['created'] = $fragment['created'] ?? null;
-
-        $item['text'] = '';
-        $item['recommendation'] = '';
-        // we need to fetch Fragment, as text fields are not mapped in statement
-        // index for performance reasons
-        $statementFragment = $this->statementHandler->getStatementFragment($fragment['id']);
-        if ($statementFragment instanceof StatementFragment) {
-            // pretend as if consideration would be an recommendation
-            // as it has the same behaviour
-            $item['recommendation'] = $statementFragment->getConsideration();
-            $item['text'] = $statementFragment->getText();
-        }
-
-        $item['elementId'] = $tmpElementId;
-        $item['elementTitle'] = $tmpElementTitle;
-
-        return $item;
     }
 
     /**
@@ -1763,48 +1496,6 @@ class DocxExporter
             }
             $cell2->addText($this->mapService->getReplacedMapAttribution($statement->getProcedure()));
         }
-    }
-
-    /**
-     * Generate Html imagetag to be used in PhpWord Html::addHtml().
-     * File needs to be locally accessible.
-     *
-     * @param string $imageFile
-     * @param int    $maxWidth  maximum image width in pixel
-     *
-     * @return string
-     */
-    protected function getDocxImageTag($imageFile, $maxWidth = 500)
-    {
-        $imgTag = '';
-        $width = 300;
-        $height = 300;
-        $margin = 10;
-        // phpword needs a local file, no need for flysystem
-        if (!file_exists($imageFile)) {
-            return $imgTag;
-        }
-
-        // get Image size
-        $imageInfo = getimagesize($imageFile);
-        if (2 < (is_countable($imageInfo) ? count($imageInfo) : 0)) {
-            $width = $imageInfo[0] - $margin;
-            $height = $imageInfo[1] - $margin;
-        }
-
-        // check that picture is not wider than allowed
-        if ($width > $maxWidth) {
-            $factor = $width / $maxWidth;
-
-            // resize Image
-            if (0 != $factor) {
-                $width /= $factor;
-                $height /= $factor;
-            }
-            $this->getLogger()->info('Docx Image resize to width: '.$width.' and height: '.$height);
-        }
-
-        return '<img height="'.$height.'" width="'.$width.'" src="'.$imageFile.'"/>';
     }
 
     /**
